Fix error handling when configure offline chat via Web UI
Show error message and don't change enable/disable state on Web UI
when API call to update offline chat settings fails

Previously the underlying meethods to load the offline chat model was
eating up the exception. So the API controller didn't know when the
chat model had failed to load. Now we bubble up the exception to let
the controller handle it