Track generic tool use in conversation logs and research iterations
Previously we had tool specific context like code and online context.

But we need a generic tool context tracker as the number and type of
tools available to Khoj expands and can be maintained by admins and
users.

This changes adds tracking this generic tool use context in a separate
ToolContext that only required a URI and content strings. The URI is
to reference/identify the tool used and content is to track the tool
results.